+++ cmd/szero/cmd_manpage.go
