+++ cmd/szero/cmd_upscale.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"github.com/jadolg/szero/pkg"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -12,32 +13,32 @@ var upCmd = &cobra.Command{
 	Example: "szero up -n default -n klum",
 	Aliases: []string{"upscale"},
 	Run: func(cmd *cobra.Command, args []string) {
-		clientset, err := getClientset(kubeconfig, kubecontext)
+		clientset, err := pkg.GetClientset(kubeconfig, kubecontext)
 		if err != nil {
 			log.Fatal(err)
 		}
 
 		ctx := context.Background()
 		for _, namespace := range namespaces {
-			deployments, err := getDeployments(ctx, clientset, namespace)
+			deployments, err := pkg.GetDeployments(ctx, clientset, namespace)
 			if err != nil {
 				log.Fatal(err)
 			}
 
 			log.Infof("Found %d deployments in namespace %s", len(deployments.Items), namespace)
-			upscaledDeployments, err := upscaleDeployments(ctx, clientset, deployments)
+			upscaledDeployments, err := pkg.UpscaleDeployments(ctx, clientset, deployments)
 			if err != nil {
 				log.Fatal(err)
 			}
 			log.Infof("Upscaled %d deployments", upscaledDeployments)
 
-			statefulsets, err := getStatefulSets(ctx, clientset, namespace)
+			statefulsets, err := pkg.GetStatefulSets(ctx, clientset, namespace)
 			if err != nil {
 				log.Fatal(err)
 			}
 
 			log.Infof("Found %d statefulsets in namespace %s", len(statefulsets.Items), namespace)
-			upscaledStatefulsets, err := upscaleStatefulSets(ctx, clientset, statefulsets)
+			upscaledStatefulsets, err := pkg.UpscaleStatefulSets(ctx, clientset, statefulsets)
 			if err != nil {
 				log.Fatal(err)
 			}
