+++ cmd/szero/cmd_version.go
