+++ pkg/deployments_test.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"context"
@@ -10,44 +10,44 @@ import (
 	"testing"
 )
 
-func TestGetStatefulsets(t *testing.T) {
+func TestGetDeployments(t *testing.T) {
 	ctx := context.Background()
 	clientset := testclient.NewClientset()
-	statefulsets, err := getStatefulSets(ctx, clientset, "default")
+	deployments, err := GetDeployments(ctx, clientset, "default")
 	assert.NoError(t, err)
-	assert.Len(t, statefulsets.Items, 0)
+	assert.Len(t, deployments.Items, 0)
 
-	statefulset := v1.StatefulSet{
+	deployment := v1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "test",
 			Namespace: "default",
 		},
 	}
-	_, err = clientset.AppsV1().StatefulSets("default").Create(ctx, &statefulset, metav1.CreateOptions{})
+	_, err = clientset.AppsV1().Deployments("default").Create(ctx, &deployment, metav1.CreateOptions{})
 	assert.NoError(t, err)
 
-	newStatefulsets, err := getStatefulSets(ctx, clientset, "default")
+	newDeployments, err := GetDeployments(ctx, clientset, "default")
 	assert.NoError(t, err)
-	assert.Len(t, newStatefulsets.Items, 1)
+	assert.Len(t, newDeployments.Items, 1)
 }
 
-func TestDownscaleStatefulSets(t *testing.T) {
+func TestDownscaleDeployments(t *testing.T) {
 	testCases := []struct {
 		name               string
-		statefulset        v1.StatefulSet
+		deployment         v1.Deployment
 		expectedDownscaled int
 		expectedReplicas   int32
 		expectedOldScale   string
 	}{
 		{
-			name: "When the statefulset was not previously downscaled then it is downscaled",
-			statefulset: v1.StatefulSet{
+			name: "When the deployment was not previously downscaled then it is downscaled",
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:        "test",
 					Namespace:   "default",
 					Annotations: map[string]string{},
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(2),
 				},
 			},
@@ -56,8 +56,8 @@ func TestDownscaleStatefulSets(t *testing.T) {
 			expectedOldScale:   "2",
 		},
 		{
-			name: "When the statefulset was previously downscaled then nothing happens",
-			statefulset: v1.StatefulSet{
+			name: "When the deployment was previously downscaled then nothing happens",
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
@@ -65,7 +65,7 @@ func TestDownscaleStatefulSets(t *testing.T) {
 						replicasAnnotation: "2",
 					},
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(0),
 				},
 			},
@@ -74,8 +74,8 @@ func TestDownscaleStatefulSets(t *testing.T) {
 			expectedOldScale:   "2",
 		},
 		{
-			name: "When the statefulset has the downscaled annotation but the replicas are not 0 then the statefulset gets downscaled",
-			statefulset: v1.StatefulSet{
+			name: "When the deployment has the downscaled annotation but the replicas are not 0 then the deployment gets downscaled",
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
@@ -83,7 +83,7 @@ func TestDownscaleStatefulSets(t *testing.T) {
 						replicasAnnotation: "2",
 					},
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(1),
 				},
 			},
@@ -98,17 +98,17 @@ func TestDownscaleStatefulSets(t *testing.T) {
 			ctx := context.Background()
 			clientset := testclient.NewClientset()
 
-			_, err := clientset.AppsV1().StatefulSets("default").Create(ctx, &tc.statefulset, metav1.CreateOptions{})
+			_, err := clientset.AppsV1().Deployments("default").Create(ctx, &tc.deployment, metav1.CreateOptions{})
 			assert.NoError(t, err)
 
-			statefulsets, err := getStatefulSets(ctx, clientset, "default")
+			deployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
-			downscaled, err := downscaleStatefulSets(ctx, clientset, statefulsets)
+			downscaled, err := DownscaleDeployments(ctx, clientset, deployments)
 			assert.NoError(t, err)
 			assert.Equal(t, tc.expectedDownscaled, downscaled)
 
-			newDeployments, err := getStatefulSets(ctx, clientset, "default")
+			newDeployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
 			for _, d := range newDeployments.Items {
@@ -121,16 +121,16 @@ func TestDownscaleStatefulSets(t *testing.T) {
 	}
 }
 
-func TestUpscaleStatefulSets(t *testing.T) {
+func TestUpscaleDeployments(t *testing.T) {
 	testCases := []struct {
 		name             string
-		statefulset      v1.StatefulSet
+		deployment       v1.Deployment
 		expectedUpscaled int
 		expectedReplicas int32
 	}{
 		{
 			name: "When the service was previously downscaled then it is upscaled",
-			statefulset: v1.StatefulSet{
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
@@ -138,7 +138,7 @@ func TestUpscaleStatefulSets(t *testing.T) {
 						replicasAnnotation: "2",
 					},
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(0),
 				},
 			},
@@ -147,12 +147,12 @@ func TestUpscaleStatefulSets(t *testing.T) {
 		},
 		{
 			name: "When the service was previously was not downscaled then nothing happens",
-			statefulset: v1.StatefulSet{
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(1),
 				},
 			},
@@ -166,17 +166,17 @@ func TestUpscaleStatefulSets(t *testing.T) {
 			ctx := context.Background()
 			clientset := testclient.NewClientset()
 
-			_, err := clientset.AppsV1().StatefulSets("default").Create(ctx, &tc.statefulset, metav1.CreateOptions{})
+			_, err := clientset.AppsV1().Deployments("default").Create(ctx, &tc.deployment, metav1.CreateOptions{})
 			assert.NoError(t, err)
 
-			statefulsets, err := getStatefulSets(ctx, clientset, "default")
+			deployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
-			upscaled, err := upscaleStatefulSets(ctx, clientset, statefulsets)
+			upscaled, err := UpscaleDeployments(ctx, clientset, deployments)
 			assert.NoError(t, err)
 			assert.Equal(t, tc.expectedUpscaled, upscaled)
 
-			newDeployments, err := getStatefulSets(ctx, clientset, "default")
+			newDeployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
 			for _, d := range newDeployments.Items {
@@ -188,19 +188,24 @@ func TestUpscaleStatefulSets(t *testing.T) {
 	}
 }
 
-func TestRestartStatefulSets(t *testing.T) {
+func int32Ptr(i int) *int32 {
+	ptr := int32(i)
+	return &ptr
+}
+
+func TestRestartDeployments(t *testing.T) {
 	testCases := []struct {
-		name        string
-		statefulset v1.StatefulSet
+		name       string
+		deployment v1.Deployment
 	}{
 		{
 			name: "When annotations are present then annotations are updated",
-			statefulset: v1.StatefulSet{
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(1),
 					Template: corev1.PodTemplateSpec{
 						ObjectMeta: metav1.ObjectMeta{
@@ -212,12 +217,12 @@ func TestRestartStatefulSets(t *testing.T) {
 		},
 		{
 			name: "When annotations are not present then annotations are initialized and then updated",
-			statefulset: v1.StatefulSet{
+			deployment: v1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default",
 				},
-				Spec: v1.StatefulSetSpec{
+				Spec: v1.DeploymentSpec{
 					Replicas: int32Ptr(1),
 				},
 			},
@@ -229,17 +234,17 @@ func TestRestartStatefulSets(t *testing.T) {
 			ctx := context.Background()
 			clientset := testclient.NewClientset()
 
-			_, err := clientset.AppsV1().StatefulSets("default").Create(ctx, &tc.statefulset, metav1.CreateOptions{})
+			_, err := clientset.AppsV1().Deployments("default").Create(ctx, &tc.deployment, metav1.CreateOptions{})
 			assert.NoError(t, err)
 
-			statefulsets, err := getStatefulSets(ctx, clientset, "default")
+			deployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
-			upscaled, err := restartStatefulSets(ctx, clientset, statefulsets)
+			upscaled, err := RestartDeployments(ctx, clientset, deployments)
 			assert.NoError(t, err)
 			assert.Equal(t, 1, upscaled)
 
-			newDeployments, err := getStatefulSets(ctx, clientset, "default")
+			newDeployments, err := GetDeployments(ctx, clientset, "default")
 			assert.NoError(t, err)
 
 			for _, d := range newDeployments.Items {
